worker-python/requirements.txt
# Web scraping
scrapling[fetchers]==0.4.1

# Web server
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Database
psycopg2-binary==2.9.10

# LLM extraction (Groq primary + OpenAI fallback)
groq==0.25.0
openai>=1.30.0

# HTML parsing (fallback + images)
beautifulsoup4==4.13.3
lxml>=6.0.2

# Env
python-dotenv==1.1.0
